Centralise condition for enabling PCH
Instead of repeating the same logic in several places, it's better to
define the condition to enable pre-compiled headers at the top-level of
our cmake infrastructure and just check the flag elsewhere. Moreover,
the behavior is not exposed as an option to users, who can opt out of it
if required.

Together with this change we need to disable PCH support during our
export_compile_commands CI job, as clang-tidy doesn't deal well with it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>